python/requirements.txt
pyTelegramBotAPI == 4.13.0
python-dotenv == 1.0.0
openai == 0.27.9